requirements.txt
asgiref==3.7.2
cfgv==3.3.1
click==8.1.6
distlib==0.3.7
Django==3.2
django-environ==0.10.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
et-xmlfile==1.1.0
filelock==3.12.2
identify==2.5.26
mccabe==0.7.0
nodeenv==1.8.0
openpyxl==3.1.2
packaging==23.1
pathspec==0.11.2
platformdirs==3.10.0
psycopg2==2.9.7
pycodestyle==2.11.0
PyJWT==2.8.0
pytz==2023.3
PyYAML==6.0.1
simplejson==3.19.1
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.7.1
virtualenv==20.24.2
